Modularization and merge of frontend and backend code
This is going to be a hybrid application, so there is a little difference between a frontend and backend code. Now all the code is in scripts/ directory.

The scripts/index is a backend application entry point (see main property of packag.json).

The scripts/frontend is the entry point of browserify bundle.

Everything else is shared.