Server side rendering with getPosts action
A new action is introduced (getPosts). It is async. The posts store is listening to it. When it is invoked, store triggers the complete child action, passing it a list of posts. This list is fake ATM.

The getPosts action is invoked by Application component's fetch method.

Routing logic of the server looks for fetch static method of the components. If such is found, it triggers it. The method have to return a promise. Data that the promise resolves to is passed to the component as props.route.data.