Get posts works on client side as well
The store logic for fetching data is separated from action handler. It is placed in a fetch method.

Action handler calls the fetch method, which returns a promise. When this promise is resolved, the data from it is passed to getPosts.completed action. The store also triggers with this data. This is what enables it on client side.

There is a button in Application component that calls the action.