Initial state gets passed from backend to frontend
So that components get hydrated when react kicks in.

On backend the state for each fetching component is stored in an object with property names matching component's route's path, e.g. '/' for application component.

When all promises are resolved, the this state object is baked as JSON into rendered html (via simple string replace).

Frontend script reads the JSON and puts it into global namespace (window.state).

The Application component looks for it's initial state first in props (client side), then in window.state.